requirements.txt
flask
camelot-py[cv]
ghostscript
pypdf2==2.2